# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Swagger Demo App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /mysql-api:
    # binds a127 app logic to a route
    x-swagger-router-controller: test_controller
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: testMysql
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /mongo-api:
    x-swagger-router-controller: test_controller
    get:
      description: Returns 'Hello' to the caller
      operationId: testMongo
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /redis-api:
    x-swagger-router-controller: test_controller
    get:
      description: Returns 'Hello' to the caller
      operationId: testRedis
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /image-upload:
    x-swagger-router-controller: test_controller
    post:
      description: Returns 'File uploaded'
      operationId: imageUpload
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: upfile
          type: file
          description: The file to upload.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authenticate/token:
    # binds a127 app logic to a route
    x-swagger-router-controller: test_controller
    post:
      description: Returns access token.
      # used as the method name of the controller
      operationId: getToken
      parameters:
        - name: body
          required: true
          in: body
          schema:
            $ref: '#/definitions/authenticateTokenRequest'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AuthTokenResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authenticate/test:
    x-swagger-router-controller: test_controller
    get:
      description: Test token
      # used as the method name of the controller
      operationId: testToken
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  authenticateTokenRequest:
    required:
      - username
      - password
    properties:
      username:
        description: The username of user.
        type: string
      password:
        description: A password of user.
        type: string
  AuthTokenResponse:
    required:
      - token
      - expires_in
    properties:
      token:
        type: string
        description: The token used for granting access.
      expires_in:
        type: integer
        format: int64
        description: Number of seconds before the token expires.
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string